Refactored code, deleted some unnecessary files and started writing BPTT example. Furthermore added first unit tests.
