+++ .gitignore
@@ -1,2 +1,3 @@
 # Folders to ignore
 **/dist/
+**/__pycache__/
