+++ src/snnax.egg-info/SOURCES.txt
@@ -1,9 +1,11 @@
+.gitignore
 AUTHORS.rst
 CHANGELOG.rst
 CONTRIBUTING.rst
 LICENSE.txt
 README.rst
 pyproject.toml
+requirements.txt
 setup.cfg
 setup.py
 tox.ini
@@ -18,7 +20,6 @@ docs/readme.rst
 docs/requirements.txt
 docs/_static/.gitignore
 examples/gestures_bptt.py
-examples/randman_bptt.py
 examples/snnax_train_dvsgesture_multiloss.py
 examples/snnax_train_nmnist_mlp_multiloss.py
 examples/utils.py
@@ -70,7 +71,7 @@ src/snnax/snn/layers/flatten.py
 src/snnax/snn/layers/iaf.py
 src/snnax/snn/layers/li.py
 src/snnax/snn/layers/lif.py
-src/snnax/snn/layers/maxpool2d.py
+src/snnax/snn/layers/maxpool.py
 src/snnax/snn/layers/srm.py
 src/snnax/snn/layers/stateful.py
 src/snnax/utils/__init__.py
