--- src/snnax/functional/__init__.py
@@ -1,1 +1,0 @@
-from .functional import *
