+++ src/snnax/snn/__init__.py
@@ -1,4 +1,3 @@
-from . import init
 from .architecture import StatefulModel, GraphStructure
 from .composed import Sequential
 from .layers.stateful import StatefulLayer
@@ -6,5 +5,5 @@ from .layers.li import LI
 from .layers.lif import SimpleLIF, LIF, LIFSoftReset, AdaptiveLIF
 from .layers.iaf import SimpleIAF, IAF
 from .layers.flatten import Flatten
-from .layers.maxpool2d import SpikingMaxPool2d
+from .layers.maxpool import SpikingMaxPool2d
