+++ src/snnax/snn/layers/maxpool.py
